requirements.txt
numpy
pandas
msgpack-python
spacy
allennlp
torch
